Initial commit.
Generic graph structure is implemented.

The Floyd-Warshall algorithm and the
Dijkstra algorithm are implemented, tested,
and benchmarked.

The algorithms solve four problems:
1. Find the minimum distance between two nodes
2. Find one of the minimum paths between two
nodes
3. Find all the minimum distances between all
nodes
4. Find one of the minimum paths between all
nodes

The Floyd-Warshall algorithm solves problem
3 and 4. Problems 1 and 2 are solved without
any further optimization, e.g. distances
between all nodes are computed, and only the
one needed is returned to solve problem 1.

Similarly, and symmetrically, Dijkstra's
algorithm solves problem 1 and 2, and
problems 3 and 4 are solved without any
further optimization, e.g. distances between
all nodes are computed recursively using
Dijkstra's algorithm. It gives all distances
having one specific origin point, and we use
it for every origin point.

Testing is done using two hand-crafted
graphs.

Benchmarks are done with minimal iterations
and forks. The goal was to generate benchmark
resources locally, which takes some time.

Hence, the results have to be taken with a
grain of salt.

At first, they generate graphs and saves them
as huge files, not shared on git, and the
goal was to have those files generated
locally. The low iterations and not ideal
conditions in which the benchmarks were done
imply a high variance and low reliability in
the results.

The true benchmarks will take even longer and
will be done in the future.